// src/models/StoreSearchQuery.ts
/*
 * Copyright (c) 2021 The Trébol eCommerce Project
 *
 * This software is released under the MIT License.
 * https://opensource.org/licenses/MIT
 */

export class ProductSearchQuery {
  nameLike?: string;
  categoryCode?: string;
  // tags?: string;
  // brand?: string;
  // brandLike?: string;
}